+++ services/motion/motion.go
@@ -47,8 +47,8 @@ type Service interface {
 		heading float64,
 		movementSensorName resource.Name,
 		obstacles []*referenceframe.GeoObstacle,
-		linearVelocity float32,
-		angularVelocity float32,
+		linearVelocity float64,
+		angularVelocity float64,
 		extra map[string]interface{},
 	) (bool, error)
 	MoveSingleComponent(
